Cast pointers for debug output.

+++ sys/arm64/arm64/gicv3_its.c
@@ -1466,7 +1466,7 @@ gicv3_its_map_msi(device_t dev, device_t child, struct intr_irqsrc *isrc,
 	girq = (struct gicv3_its_irqsrc *)isrc;
 
 	/* For debug purposed only! */
-	device_printf(child, "%s: isrc=%#jx\n", __func__, isrc);
+	device_printf(child, "%s: isrc=%#jx\n", __func__, (uintmax_t)isrc);
 
 	*addr = vtophys(rman_get_virtual(sc->sc_its_res)) + GITS_TRANSLATER;
 	*data = girq->gi_id;
